refactor: ObjectMapper 필드로 변경
